tests/wpt/web-platform-tests/workers/interfaces/WorkerUtils/importScripts/004.html

<!doctype html>
<title>importScripts broken script</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
async_test(function() {
  var worker = new Worker('004.js');
  worker.onmessage = this.step_func(function(e) {
    assert_equals(e.data[0], "first script successful. ");
    assert_true(e.data[1], 'expected SyntaxError');
    this.done();
  });
});
</script>
